Provide connected components
Partially addressing https://github.com/proteneer/timemachine/pull/399#discussion_r618763078
